remove unused folers.
