+++ ui/src/components/AssetDetails/AssetDetails.tsx
@@ -1,4 +1,4 @@
-import { Button, Divider, Typography } from '@mui/material';
+import {  Divider, Typography } from '@mui/material';
 import React from 'react';
 import { Theme } from '@mui/material/styles';
 import { makeStyles } from '@mui/styles';
