+++ ui/src/components/PendingAssetInviteRowContent/PendingAssetInviteRowContent.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import Typography from '@mui/material/Typography';
+import {  Divider } from '@mui/material';
+
+import clx from 'clsx'
+import { useNarrowPendingStyles } from '../PendingRowContents/PendingRowContents';
+import { PendingRowProps } from '../PendingRow/PendingRow';
+
+export const PendingAssetInviteRowContent: React.FC<PendingRowProps> = ({ isInbound, sender, inboundTicker }) => {
+  const classes = useNarrowPendingStyles();
+
+  const inboundMessage = (
+    <>
+      <div className={classes.divider} />
+      <Typography variant='body2' className={classes.text} color="text.primary" >
+        {sender}
+      </Typography>
+      <Divider className={classes.divider} />
+      <Typography variant='body2' color='text.secondary' className={classes.text} >
+        Invites you to create asset holding account for
+          </Typography>
+      <Divider className={classes.divider} />
+
+      <Typography variant='body2' className={clx(classes.text, classes.assetName)} color="text.secondary" >
+        {inboundTicker}
+      </Typography>
+      <div className={classes.divider} />
+
+    </>
+  )
+  const outboundMessage = (
+    <>
+      <div className={classes.divider} />
+      <Typography variant='body2' color='text.secondary' className={classes.text} >
+        You are inviting {sender} to create asset holding account for
+    </Typography>
+      <Divider className={classes.divider} />
+      <Typography variant='body2' className={clx(classes.text, classes.assetName)} color="text.secondary" >
+        {inboundTicker}
+      </Typography>
+      <div className={classes.divider} />
+    </>
+  )
+
+  return (
+    <>
+          <div>
+            {isInbound ? inboundMessage : outboundMessage}
+          </div>
+    </>
+  );
+}
+
+
