+++ ui/src/pages/SendPage.tsx
@@ -1,15 +1,10 @@
 import React from 'react';
-import { SendPopupContent } from '../components/SendPopupContent/SendPopupContent';
-import { useNavigate, useLocation, useParams } from 'react-router-dom'
-import { Theme } from '@mui/material/styles';
-import { makeStyles } from '@mui/styles';
+import { useNavigate, useParams } from 'react-router-dom'
 import ArrowBackIosNewIcon from '@mui/icons-material/ArrowBackIosNew';
 import { Avatar, Box, Card, CardContent, IconButton, Typography } from '@mui/material';
 import { SendForm } from '../components/SendForm/SendForm';
 import { isMobile } from '../platform/platform';
 import { usePageStyles } from './AssetProfilePage';
-
-
 
 export const SendPage: React.FC = () => {
   const nav = useNavigate();
